feat(Carrera): Add payment functions
Added the realizarPago, getter and setter for propina and liberarConductor methods.

+++ src/main/java/edu/estatuas/cotxox/conductores/Conductor.java
@@ -53,7 +53,7 @@ public class Conductor {
 
     private double calcularValoracionMedia() {
 
-        int sumValoraciones = 0;
+        double sumValoraciones = 0;
         for (Byte valoracion : valoraciones) {
             sumValoraciones += valoracion;
         }
